requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
scikit-learn